+++ lib/amoeba/cloner.rb
@@ -127,7 +127,9 @@ module Amoeba
 
     def process_null_fields
       # nullify any fields the user has configured
-      amoeba.null_fields.each do |field_key|
+      amoeba.null_fields.each do |field_key, options|
+        next if options[:if] && !@old_object.send(options[:if])
+
         @new_object[field_key] = nil
       end
     end